pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-genmedia-mcp"
version = "0.1.0"
description = "MCP server for Google GenMedia APIs (Imagen, Gemini Image, Veo, Chirp, Lyria)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "google-genai>=1.0.0",
    "google-cloud-storage>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
phase2 = [
    "google-cloud-texttospeech>=2.0.0",
    "google-cloud-aiplatform>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
google-genmedia-mcp = "google_genmedia_mcp.mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/google_genmedia_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
# 日本語文字列リテラルを含む行は行長制限を超えることがあるため除外
extend-ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
# サードパーティ製ライブラリのスタブが不完全なものは無視
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
